3d topology feature gate
