+++ web/src/components/netflow-topology/2d/componentFactories/shapesComponentFactory.ts
@@ -1,6 +1,6 @@
 import { ComponentType } from 'react';
 import { ComponentFactory, ModelKind, DefaultNode } from '@patternfly/react-topology';
-import { GraphElementPeer } from '../../../model/topology';
+import { GraphElementPeer } from '../../../../model/topology';
 
 export const shapesComponentFactory: ComponentFactory = (
   kind: ModelKind,
