+++ web/src/components/netflow-topology/2d/components/edge.tsx
@@ -23,7 +23,7 @@ import {
 import DefaultConnectorTag from '@patternfly/react-topology/dist/esm/components/edges/DefaultConnectorTag';
 import { getConnectorStartPoint } from '@patternfly/react-topology/dist/esm/components/edges/terminals/terminalUtils';
 
-import { HOVER_EVENT } from '../netflow-topology';
+import { HOVER_EVENT } from '../topology-content';
 
 type BaseEdgeProps = {
   children?: React.ReactNode;
