+++ web/src/components/netflow-topology/2d/layouts/layoutFactory.ts
@@ -10,7 +10,7 @@ import {
   BreadthFirstLayout,
   ColaGroupsLayout
 } from '@patternfly/react-topology';
-import { LayoutName } from '../../../model/topology';
+import { LayoutName } from '../../../../model/topology';
 
 const layoutFactory: LayoutFactory = (type: LayoutName, graph: Graph): Layout | undefined => {
   switch (type) {
